#825 Support checking palette last committed time before performing updates
